refactor : Combine service class
