+++ wise-office-server/src/main/java/kr/co/wise/office/domain/minutes/service/MinutesService.java
@@ -12,21 +12,24 @@ import kr.co.wise.office.domain.minutes.repository.MinutesEntityRepository;
 import kr.co.wise.office.domain.minutesattendant.repository.MinutesAttendantEntityRepository;
 import kr.co.wise.office.domain.minutesattendant.repository.MinutesAttendantEntityRepository.MinutesAttendantsInfoProjection;
 import kr.co.wise.office.domain.proposalattendant.entity.ProposalAttendantEntity;
-import kr.co.wise.office.domain.proposalattendant.service.ProposalAttendantsService;
+import kr.co.wise.office.domain.proposalattendant.service.ProposalAttendantService;
 import kr.co.wise.office.exception.ErrorMessage;
 import kr.co.wise.office.exception.custom.NotFoundResourceException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.data.domain.Pageable;
+
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.stream.Collectors;
+
+;
 
 @Service
 @RequiredArgsConstructor
@@ -39,7 +42,8 @@ public class MinutesService {
 
     private final ApproveEntityRepository approveEntityRepository;
 
-    private final ProposalAttendantsService proposalAttendantsService;
+    private final ProposalAttendantService proposalAttendantsService;
+
     public List<MinutesListResponse> getMinutesBriefInfo(
             long projectId
     ) {
