+++ wise-office-server/src/main/java/kr/co/wise/office/domain/proposalattendant/service/ProposalAttendantService.java
@@ -5,6 +5,8 @@ import kr.co.wise.office.domain.companymember.entity.CompanyMemberEntity;
 import kr.co.wise.office.domain.proposalattendant.entity.ProposalAttendantEntity;
 import kr.co.wise.office.domain.proposalattendant.entity.ProposalAttendantRole;
 import kr.co.wise.office.domain.proposalattendant.repository.ProposalAttendantEntityRepository;
+import kr.co.wise.office.exception.ErrorMessage;
+import kr.co.wise.office.exception.custom.NotFoundResourceException;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -82,4 +84,34 @@ public class ProposalAttendantService {
         // 배치 저장
         proposalAttendantEntityRepository.saveAll(newAttendants);
     }
+
+    public List<ProposalAttendantEntity> findByProjectId(long projectId) {
+        //제안서상 편성 인원 조회
+        return proposalAttendantEntityRepository.findByProjectIdWithCompanyName(projectId);
+    }
+
+    public ProposalAttendantEntity findWriterInfo(long writerProposalAttendantId, long projectId) {
+        return proposalAttendantEntityRepository.findWriterInfo(writerProposalAttendantId, projectId)
+                .orElseThrow(() -> new NotFoundResourceException(ErrorMessage.REJECT_CREATE_MINUTES));
+    }
+
+    public List<ProposalAttendantEntity> findWriterInfos(List<Long> proposalIds, long projectId) {
+        if (proposalIds.isEmpty()) {
+            return List.of();
+        }
+        List<ProposalAttendantEntity> writerInfos = proposalAttendantEntityRepository.findWriterInfos(proposalIds, projectId);
+        if (writerInfos.isEmpty()) {
+            throw new NotFoundResourceException(ErrorMessage.NOT_FOUND_ATTENDANT);
+        }
+        return writerInfos;
+    }
+
+    // 모든 프로젝트 탈퇴처리
+    @Transactional
+    public void exitAllProject(List<CompanyMemberEntity> exitCompanyMembers) {
+        List<ProposalAttendantEntity> exitProposalAttendants = proposalAttendantEntityRepository.findByCompanyMemberInAndExitDateIsNull(exitCompanyMembers);
+        exitProposalAttendants.forEach(ProposalAttendantEntity::leaveProject);
+        proposalAttendantEntityRepository.saveAll(exitProposalAttendants);
+    }
+
 }
